src/main/java/org/han/ica/oose/boterbloem/domain/Zorginstelling.java
package org.han.ica.oose.boterbloem.domain;

/**
 * Representative domain class for a Zorginstelling
 */
public class Zorginstelling {
    private int id;
    private String name;

    public Zorginstelling(int id, String name) {
        this.id = id;
        this.name = name;
    }

    /**
     * Getter for property 'id'.
     *
     * @return Value for property 'id'.
     */
    public int getId() {
        return id;
    }

    /**
     * Setter for property 'id'.
     *
     * @param id Value to set for property 'id'.
     */
    public void setId(int id) {
        this.id = id;
    }

    /**
     * Getter for property 'name'.
     *
     * @return Value for property 'name'.
     */
    public String getName() {
        return name;
    }

    /**
     * Setter for property 'name'.
     *
     * @param name Value to set for property 'name'.
     */
    public void setName(String name) {
        this.name = name;
    }
}
